requirements.txt
tensorflow
numpy
pillow
matplotlib
flask
scipy
opencv-python-headless
